--- FileTemplate/templates/basicHtml.html
@@ -1,13 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <head>
-        <meta charset="UTF-8">
-        <meta name="viewport" content="width = device-width, initial-scale = 1.0">
-        <meta http-equiv="X-UA-Compatible" content = "ie=edge">
-        <title> Keat </title>
-    </head>
-    
-    <body>
-        
-    </body>
-</html>
