--- FileTemplate/templates/basicphp.php
@@ -1,4 +1,0 @@
-<?php
-    $q = $_REQUEST['q'];
-    echo "$q";
-?>
