+++ file-templates/templates/basicVue.html
@@ -1,0 +1,27 @@
+<!DOCTYPE html>
+<html>
+    <head>
+        <meta charset="UTF-8">
+        <meta name="viewport" content="width = device-width, initial-scale = 1.0">
+        <meta http-equiv="X-UA-Compatible" content = "ie=edge">
+        <script src="https://cdn.jsdelivr.net/npm/vue/dist/vue.js"></script>
+        <title> Keat </title>
+    </head>
+    
+    <body>
+        
+        <div id="app">
+            {{ message }}
+        </div>
+        
+        <script>
+            var app = new Vue({
+                el: '#app',
+                data: {
+                message: 'Hello Vue!'
+                }
+            })
+        </script>
+        
+    </body>
+</html>
