+++ file-templates/templates/basicphp.php
@@ -1,0 +1,4 @@
+<?php
+    $q = $_REQUEST['q'];
+    echo "$q";
+?>
